converter-script/requirements.txt

python-pptx==0.6.23
requests==2.31.0
beautifulsoup4==4.12.3
Pillow==10.4.0
pika==1.3.2
redis==5.0.1
pylint==3.0.3

